app/requirements.txt
psycopg2-binary
fastapi
uvicorn
celery
redis
flower